+++ backend/src/main/java/backend/mulkkam/averageTemperature/serivce/ScheduleService.java
@@ -1,6 +1,7 @@
 package backend.mulkkam.averageTemperature.serivce;
 
 import backend.mulkkam.intake.service.WeatherService;
+import backend.mulkkam.intakenotification.IntakeNotificationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -10,6 +11,7 @@ import org.springframework.stereotype.Component;
 public class ScheduleService {
 
     private final WeatherService weatherService;
+    private final IntakeNotificationService intakeNotificationService;
 
     @Scheduled(cron = "0 0 19 * * *")
     public void saveTomorrowAverageTemperature() {
@@ -18,6 +20,6 @@ public class ScheduleService {
 
     @Scheduled(cron = "0 0 8 * * *")
     public void notifyAdditionalWaterIntakeByWeather() {
-        weatherService.notifyAdditionalIntakeByStoredWeather();
+        intakeNotificationService.notifyAdditionalIntakeByStoredWeather();
     }
 }
