+++ backend/src/main/java/backend/mulkkam/weather/WeatherController.java
@@ -1,6 +1,6 @@
 package backend.mulkkam.weather;
 
-import backend.mulkkam.intake.service.WeatherService;
+import backend.mulkkam.intakenotification.IntakeNotificationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -12,11 +12,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/weather")
 public class WeatherController {
 
-    private final WeatherService weatherService;
+    private final IntakeNotificationService intakeNotificationService;
 
     @PostMapping
     public ResponseEntity<Void> create() {
-        weatherService.notifyAdditionalIntakeByStoredWeather();
+        intakeNotificationService.notifyAdditionalIntakeByStoredWeather();
         return ResponseEntity.ok().build();
     }
 }
